requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
boto3==1.40.12
python-multipart==0.0.20
httpx==0.28.1
Pillow==12.1.0
